Refactor SSH workflow functions for clarity
- Consolidate variable definitions and initialization to simplify logic.
- Streamline error handling and improve message visibility during execution.
- Enhance modularity by introducing `do_ssh_export_all` for exporting SSH variables.

+++ workflow.ssh.sh
@@ -5,19 +5,20 @@ set -eu -o pipefail
 
 define_ssh() {
   declare -F 'ssh_prepare' >/dev/null && return 0
-  SSH_COMMAND="${SSH_COMMAND:-${SSH_PROXY_JUMP:-}}"
-  SSH_WORK_HOME="${SSH_WORK_HOME:-$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)}"
-  SSH_NESTED_WORKFLOW="${SSH_NESTED_WORKFLOW:-none}"
-  SSH_ARCHIVE_NAME="${SSH_ARCHIVE_NAME:-${SSH_NESTED_WORKFLOW-}}"
-  SSH_ARCHIVE_PATH="${SSH_ARCHIVE_PATH:-/tmp/${SSH_WORK_HOME##*/}}-${SSH_ARCHIVE_NAME-}.tgz"
-  SSH_PROCESS_HOME="${SSH_PROCESS_HOME:-/tmp/${SSH_WORK_HOME##*/}}"
-  define_none() { :; }
   ssh_prepare() {
     do_print_section "SSH WORKFLOW BEGIN"
     do_print_dash_pair "${FUNCNAME[0]}"
+    SSH_COMMAND="${ARG_SSH_COMMAND:-${SSH_PROXY_JUMP:-}}"
     do_print_dash_pair 'SSH_COMMAND' "${SSH_COMMAND-}"
     [ -z "${SSH_COMMAND-}" ] && return 11
     do_ssh_ensure_agent || return $?
+    local SSH_WORK_HOME="${ARG_SSH_WORK_HOME:-$(cd "$(dirname "${BASH_SOURCE[0]}")/.." && pwd)}"
+    SSH_ARCHIVE_GROUP="${ARG_SSH_ARCHIVE_GROUP:-${SSH_WORK_HOME##*/}}"
+    SSH_ARCHIVE_DIR="${ARG_SSH_ARCHIVE_DIR-}"
+    SSH_ARCHIVE_NAME="${ARG_SSH_ARCHIVE_NAME:?}"
+    SSH_ARCHIVE_PATH="${ARG_SSH_ARCHIVE_PATH:-/tmp/${SSH_ARCHIVE_GROUP:?}/${SSH_ARCHIVE_NAME:?}.tgz}"
+    SSH_PROCESS_HOME="${ARG_SSH_PROCESS_HOME:-/tmp/${SSH_ARCHIVE_GROUP:?}}"
+    SSH_NESTED_WORKFLOW="${ARG_SSH_NESTED_WORKFLOW:-process_${SSH_ARCHIVE_NAME}}"
     _prepare_ssh_key
     _prepare_ssh_known
     do_func_invoke ssh_prepare_do
@@ -25,21 +26,28 @@ define_ssh() {
   }
   ssh_upload() {
     do_print_dash_pair "${FUNCNAME[0]}"
+    do_print_dash_pair 'SSH_ARCHIVE_DIR' "${SSH_ARCHIVE_DIR-}"
     do_print_dash_pair 'SSH_ARCHIVE_NAME' "${SSH_ARCHIVE_NAME-}"
     do_print_dash_pair 'SSH_ARCHIVE_PATH' "${SSH_ARCHIVE_PATH-}"
     [ -z "${SSH_ARCHIVE_NAME-}" ] && return 11
     [ -z "${SSH_ARCHIVE_PATH-}" ] && return 12
     "define_${SSH_NESTED_WORKFLOW-}" || return $?
-    do_ssh_archive_dir "$SSH_ARCHIVE_NAME" "$SSH_COMMAND" "$SSH_ARCHIVE_PATH" || return "$?"
+    [ -n "${SSH_ARCHIVE_DIR-}" ] && { pushd "${SSH_ARCHIVE_DIR-}" || return $?; }
+    do_ssh_archive_dir "$SSH_ARCHIVE_NAME" "$SSH_COMMAND" "$SSH_ARCHIVE_PATH" || return $?
+    popd || :
     do_print_trace "$(do_stack_trace)" done!
+  }
+  do_ssh_export_all() {
+    do_ssh_export_reset
+    do_ssh_export SSH_ARCHIVE_GROUP SSH_ARCHIVE_NAME \
+      SSH_ARCHIVE_PATH SSH_PROCESS_HOME SSH_NESTED_WORKFLOW
+    do_ssh_export define_cidoer_print
+    do_func_invoke ssh_export_do
   }
   ssh_process_finally() { return 0; }
   ssh_process() {
     do_print_dash_pair "${FUNCNAME[0]}"
-    do_ssh_export_reset
-    do_func_invoke ssh_export_do
-    do_ssh_export SSH_NESTED_WORKFLOW SSH_ARCHIVE_NAME SSH_ARCHIVE_PATH SSH_PROCESS_HOME
-    do_ssh_export define_cidoer_print
+    do_ssh_export_all
     do_ssh_exec "$SSH_COMMAND" define_cidoer_core "define_${SSH_NESTED_WORKFLOW-}" _process
     do_print_trace "$(do_stack_trace)" done!
   }
@@ -47,31 +55,44 @@ define_ssh() {
     [ -n "${SSH_ARCHIVE_NAME-}" ] || return 11
     [ -n "${SSH_ARCHIVE_PATH-}" ] || return 12
     local -r dir="${SSH_PROCESS_HOME:-/tmp}"
-    mkdir -p "$dir" || return $?
+    mkdir -p "$dir/${SSH_ARCHIVE_NAME-}" || return $?
     pushd "$dir" >/dev/null || return $?
     tar xzf "$SSH_ARCHIVE_PATH" || return $?
     popd >/dev/null
     pushd "$dir/${SSH_ARCHIVE_NAME-}" >/dev/null || return $?
+    do_print_dash_pair 'SSH_ARCHIVE_GROUP' "${SSH_ARCHIVE_GROUP-}"
+    do_print_dash_pair 'SSH_ARCHIVE_NAME' "${SSH_ARCHIVE_NAME-}"
+    do_print_dash_pair 'SSH_ARCHIVE_PATH' "${SSH_ARCHIVE_PATH-}"
+    do_print_dash_pair 'SSH_PROCESS_HOME' "${SSH_PROCESS_HOME-}"
     do_func_invoke ssh_process_do
     popd >/dev/null
   }
   ssh_prune_finally() { return 0; }
   ssh_prune() {
     do_print_dash_pair "${FUNCNAME[0]}"
+    do_ssh_export_all
     do_ssh_exec "$SSH_COMMAND" define_cidoer_core "define_${SSH_NESTED_WORKFLOW-}" _prune
     do_print_trace "$(do_stack_trace)" done!
   }
   _prune() {
-    [ -n "${SSH_ARCHIVE_NAME-}" ] || return 11
-    [ -n "${SSH_ARCHIVE_PATH-}" ] || return 12
+    do_print_dash_pair "${FUNCNAME[0]}"
+    [ -n "${SSH_ARCHIVE_PATH-}" ] || {
+      do_print_warn 'SSH_ARCHIVE_PATH is absent'
+      return 11
+    }
     rm "${SSH_ARCHIVE_PATH:-}" || :
     rm "${SSH_ARCHIVE_PATH:-}.sha256" || :
-    local -r dir="${SSH_PROCESS_HOME:-/tmp}"
-    pushd "$dir" >/dev/null || return $?
-    do_func_invoke ssh_prune_do
-    rm -rf "$SSH_ARCHIVE_NAME" || :
+    [ -n "${SSH_PROCESS_HOME-}" ] || {
+      do_print_warn 'SSH_PROCESS_HOME is absent'
+      return 12
+    }
+    pushd "$SSH_PROCESS_HOME" >/dev/null || return $?
+    do_func_invoke ssh_prune_do || :
+    [ -n "${SSH_ARCHIVE_NAME-}" ] && { rm -rf "$SSH_ARCHIVE_NAME" || :; }
     popd >/dev/null
+    do_print_trace "$(do_stack_trace)" done!
   }
+
   ssh_finish_finally() {
     do_print_section "SSH WORKFLOW DONE!"
     return 0
